Add Genius
May have a bug or two to iron out but for the most part, it works.
This requires a file 'token.txt' at root from Genius API to be used.
Make an API client at http://genius.com/api-clients and generate a
client access token, saving that value in the file.